refactor: s3 관련 기능 최적화 #4
ImageInfo를 이용하여 파일 삭제 기능 구현
